+++ docs/change_log.md
@@ -1524,3 +1524,39 @@ java.lang.NoClassDefFoundError: javax/servlet/http/HttpSession
 
 !string.startsWithAny(package_name, 'a.b.c.d1', 'a.b.c.d2')
 ```
+
+## 1.38. (3.3.2)
+
+### 1.38.1. JarDiff 功能增加配置参数
+
+- _jacg_jar_diff/jar_diff_callee_method_prefix.properties
+
+```
+# （作用） JarDIff 比较新旧两个目录的 jar 文件，获得发生变化的方法的影响范围（生成向上的完整方法调用链及调用堆栈）使用的配置文件（每行指定一项配置，可指定多行）
+# （内容） 指定发生变化的方法中，需要进行处理的方法前缀（若未指定则不限制）
+# （格式） 可指定包名，或包名+类名，或包名+类名+方法名，或包名+类名+方法名+参数
+# （示例）
+# com.test
+# com.test.Test1
+# com.test.Test1:func1
+# com.test.Test1:func1(
+# com.test.Test1:func1(java.lang.String)
+```
+
+示例见 test.runbycode.jardiffcallgraph.TestRunnerGenJarDiffCallGraphDiffSame:testJarDiffCalleeGraphChooseMethod 方法
+
+- _jacg_jar_diff/jar_diff_caller_method_prefix.properties
+
+```
+# （作用） JarDIff 比较新旧两个目录的 jar 文件，获得发生变化的方法向下的完整方法调用链使用的配置文件（每行指定一项配置，可指定多行）
+# （内容） 指定发生变化的方法中，需要进行处理的方法前缀（若未指定则不限制）
+# （格式） 可指定包名，或包名+类名，或包名+类名+方法名，或包名+类名+方法名+参数
+# （示例）
+# com.test
+# com.test.Test1
+# com.test.Test1:func1
+# com.test.Test1:func1(
+# com.test.Test1:func1(java.lang.String)
+```
+
+示例见 test.runbycode.jardiffcallgraph.TestRunnerGenJarDiffCallGraphDiffSame:testJarDiffCallerGraphChooseMethod 方法